U8 debug  Changes to be committed: 	modified:   src/ai/ai.c 	modified:   src/entities/TBall.c 	modified:   src/entities/TBall.h 	modified:   src/entities/TPlayer.c 	modified:   src/game.c 	modified:   src/levels/court01.c 	modified:   src/levels/court01.h

+++ src/entities/TBall.h
@@ -40,7 +40,7 @@ extern TBall ball;
 
 void updateBall(TBall *ball);
 void initBall(TBall *ball);
-void newBall(i32 x, i32 y, TBall *ball);
+void newBall(u8 x, u8 y, TBall *ball);
 void calcBounce(TBall *ball);
 
 #endif
